pyproject.toml
[tool.poetry]
name = "flask-api-book-store"
version = "0.1.0"
description = ""
authors = ["fauzanalghifary <m.fauzanghifary@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
flask-sqlalchemy = "^3.1.1"
flask-smorest = "^0.43.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
gunicorn = "^21.2.0"
pytest = "^8.0.1"
pytest-cov = "^4.1.0"
flask-migrate = "^4.0.5"
flask-login = "^0.6.3"
flask-jwt-extended = "^4.6.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
